feat(console): decentralize settings — live where the thing lives
Kill the monolithic Settings surface. Settings now render in their home view:
- Agent settings (model, routing, goal mode, tools) → a Settings tab in the Agent view.
- Memory settings (knowledge) → a Settings tab in the Knowledge view.
- Central Settings = only cross-cutting tabs: Overview · Telemetry · Plugins · System.
  Telemetry split out of Overview; Integrations renamed Plugins. View-less plugins'
  settings default to Settings → Plugins; a plugin with a view owns its own.

- Extract <SettingsCategory> / <SettingsCategoryPanel> from SettingsSurface — a
  self-contained per-category renderer (dirty + Save + all the Test buttons) any view
  embeds. Per-category dirty state (you save settings where they live).
- Backend: re-home sections (graph/settings_schema.py) — Goal mode/Tools → Agent,
  Compaction/Caching → System, default → Plugins; categories Agent/Memory/Plugins/System.
- App.tsx: Agent gains a Settings tab; Knowledge gains Store/Settings tabs; central
  Settings is tab-driven (Overview/Telemetry/Plugins/System).

Tests: settings_schema reordering; e2e rewritten for the new homes (Agent settings in
Agent view, telemetry/delegates relocated). Full suite 1260 + e2e 60.
