+++ apps/web/src/settings/SettingsSurface.tsx
@@ -1,440 +1,36 @@
-import { QueryErrorResetBoundary, useMutation, useQuery, useQueryClient, useSuspenseQuery } from "@tanstack/react-query";
-import { AlertTriangle, ExternalLink, Link2, Loader2, RotateCcw, Save, ShieldCheck } from "lucide-react";
-import { Suspense, useMemo, useState } from "react";
+import { BarChart3, Gauge, Puzzle, Settings2 } from "lucide-react";
 
-// Setup walkthroughs live in the template's (protoAgent) docs — forks don't ship
-// their own docs site, so the in-app help links point at the canonical docs.
-const DISCORD_GUIDE_URL = "https://protolabsai.github.io/protoAgent/guides/discord#bot-setup";
-const GOOGLE_GUIDE_URL = "https://protolabsai.github.io/protoAgent/guides/google#oauth-client";
-
-import { ErrorBoundary, PanelError, PanelSkeleton } from "../app/ErrorBoundary";
-import { StageSubnav } from "../app/StageSubnav";
-import { PanelHeader } from "../app/PanelHeader";
-import { api } from "../lib/api";
-import { queryKeys, settingsSchemaQuery } from "../lib/queries";
-import type { SettingsField, SettingsGroup } from "../lib/types";
+import { TelemetrySurface } from "../telemetry/TelemetrySurface";
 import { DelegatesSection } from "./DelegatesSection";
 import { OverviewPanel } from "./OverviewPanel";
+import { SettingsCategoryPanel } from "./SettingsCategory";
 
-// Generic settings surface — renders whatever GET /api/settings/schema returns,
-// so it stays in sync as config grows. Saving POSTs the changed fields and the
-// server hot-reloads the agent; fields flagged `restart` get a badge + banner.
-// On the TanStack Query data layer (ADR 0013): the schema is a useSuspenseQuery,
-// save is a useMutation that invalidates it; loading/errors via Suspense +
-// ErrorBoundary.
+// Central Settings — only cross-cutting stuff now (the settings decentralization):
+// Overview (status), Telemetry, Plugins (delegates + the settings for any view-less
+// plugin), and System (runtime/middleware). Agent + Memory settings moved to their
+// home views (Agent → Settings, Knowledge → Settings).
 
-export function SettingsSurface() {
-  // SettingsBody owns its own panel <section> (so the sub-tab strip can sit ABOVE
-  // the card, consistent with the rail surfaces). The loading/error fallbacks carry
-  // their own card so the layout doesn't jump.
-  return (
-    <QueryErrorResetBoundary>
-      {({ reset }) => (
-        <ErrorBoundary
-          onReset={reset}
-          fallback={(a) => (
-            <section className="panel stage-panel settings-panel">
-              <PanelError {...a} label="settings" />
-            </section>
-          )}
-        >
-          <Suspense
-            fallback={
-              <section className="panel stage-panel settings-panel">
-                <PanelSkeleton label="Loading settings…" />
-              </section>
-            }
-          >
-            <SettingsBody />
-          </Suspense>
-        </ErrorBoundary>
-      )}
-    </QueryErrorResetBoundary>
-  );
-}
+export type SettingsTab = "overview" | "telemetry" | "plugins" | "system";
 
-function SettingsBody() {
-  const queryClient = useQueryClient();
-  const { data } = useSuspenseQuery(settingsSchemaQuery());
-  const groups = data.groups;
-  const [dirty, setDirty] = useState<Record<string, unknown>>({});
-  const [status, setStatus] = useState("");
+export const SETTINGS_TABS = [
+  { id: "overview", label: "Overview", icon: Gauge },
+  { id: "telemetry", label: "Telemetry", icon: BarChart3 },
+  { id: "plugins", label: "Plugins", icon: Puzzle },
+  { id: "system", label: "System", icon: Settings2 },
+] as const;
 
-  // Category sub-nav (ADR 0020): the server tags each group with a category and
-  // orders them, so we derive the ordered category list by first appearance.
-  const categories = useMemo(() => {
-    // Overview leads — the read-only status snapshot + telemetry (moved here from
-    // the old Runtime section).
-    const seen: string[] = ["Overview"];
-    for (const g of groups) {
-      const c = g.category || "Integrations";
-      if (!seen.includes(c)) seen.push(c);
-    }
-    // Integrations always shows — the delegate registry (ADR 0025) lives here as
-    // a custom CRUD panel (plugins moved to their own top-level section).
-    if (!seen.includes("Integrations")) seen.push("Integrations");
-    return seen;
-  }, [groups]);
-  const [activeCategory, setActiveCategory] = useState(categories[0] || "");
-  // The active category must stay valid if the schema reshapes under us.
-  const category = categories.includes(activeCategory) ? activeCategory : categories[0] || "";
-  const visibleGroups = groups.filter((g) => (g.category || "Integrations") === category);
-
-  const dirtyKeys = Object.keys(dirty);
-
-  // Pending changes that won't take effect until a process restart.
-  const pendingRestart = useMemo(() => {
-    const labels: string[] = [];
-    for (const g of groups) {
-      for (const f of g.fields) {
-        if (f.restart && f.key in dirty) labels.push(f.label);
-      }
-    }
-    return labels;
-  }, [groups, dirty]);
-
-  const save = useMutation({
-    mutationFn: () => api.saveSettings(dirty),
-    onMutate: () => setStatus("saving…"),
-    onSuccess: (r) => {
-      if (!r.ok) {
-        setStatus(`save failed: ${r.messages.join(" · ")}`);
-        return;
-      }
-      const restartNote = r.restart_required.length
-        ? ` — restart required for: ${r.restart_required.join(", ")}`
-        : "";
-      setStatus(`${r.messages.join(" · ")}${restartNote}`);
-      setDirty({});
-      void queryClient.invalidateQueries({ queryKey: queryKeys.settings });
-    },
-    onError: (e) => setStatus(`save failed: ${e instanceof Error ? e.message : String(e)}`),
-  });
-
-  // Verify the model can actually complete — tests the pending edits if any
-  // (e.g. a freshly-typed key), else the saved config (blanks fall back server
-  // side). Real completion probe, the same auth path as chat.
-  const asStr = (v: unknown) => (typeof v === "string" ? v : "");
-  const testConn = useMutation({
-    mutationFn: () =>
-      api.testModel(
-        asStr(dirty["model.api_base"]),
-        asStr(dirty["model.api_key"]),
-        asStr(dirty["model.name"]),
-      ),
-    onMutate: () => setStatus("testing connection…"),
-    onSuccess: (r) =>
-      setStatus(r.ok ? "connection OK — the model responded." : `connection failed — ${r.error || "no response"}`),
-    onError: (e) => setStatus(`connection test failed: ${e instanceof Error ? e.message : String(e)}`),
-  });
-
-  // Generic "Test connection" for any group whose plugin declares one (ADR 0029).
-  // Posts the group's fields (short keys; unset secrets fall back to saved config).
-  const [testingSection, setTestingSection] = useState<string | null>(null);
-  const groupFields = (group: SettingsGroup): Record<string, unknown> => {
-    const out: Record<string, unknown> = {};
-    for (const f of group.fields) {
-      const short = f.key.split(".").pop() as string;
-      if (f.key in dirty) out[short] = dirty[f.key];
-      else if (f.type !== "secret") out[short] = f.value;  // never echo an unset secret
-    }
-    return out;
-  };
-  const testGroup = useMutation({
-    mutationFn: (vars: { endpoint: string; fields: Record<string, unknown> }) =>
-      api.testConfig(vars.endpoint, vars.fields),
-    onMutate: () => setStatus("testing connection…"),
-    onSuccess: (r) =>
-      setStatus(r.ok ? `connection OK${r.identity ? ` — ${r.identity}` : ""}` : `connection failed — ${r.error || "no response"}`),
-    onError: (e) => setStatus(`connection test failed: ${e instanceof Error ? e.message : String(e)}`),
-    onSettled: () => setTestingSection(null),
-  });
-
-  // Verify a Discord bot token (pending edit or saved). Shows the bot name.
-  const testDiscord = useMutation({
-    mutationFn: () => api.testDiscord(asStr(dirty["discord.bot_token"])),
-    onMutate: () => setStatus("testing Discord…"),
-    onSuccess: (r) =>
-      setStatus(
-        r.ok
-          ? `Discord OK — connected as ${r.bot_user || "your bot"}.`
-          : `Discord connection failed — ${r.error || "check the token"}`,
-      ),
-    onError: (e) => setStatus(`Discord test failed: ${e instanceof Error ? e.message : String(e)}`),
-  });
-
-  // Google surface (ADR 0017): show connection status + a "Connect Google"
-  // button that runs the OAuth consent (opens the operator's browser).
-  const googleStatus = useQuery({ queryKey: ["google-status"], queryFn: () => api.googleStatus() });
-  const googleConnect = useMutation({
-    mutationFn: () => api.googleConnect(),
-    onMutate: () => setStatus("opening Google consent in your browser…"),
-    onSuccess: (r) => {
-      setStatus(
-        r.ok
-          ? `Google connected${r.email ? ` as ${r.email}` : ""}.`
-          : `Google connect failed — ${r.error || "try again"}`,
-      );
-      void googleStatus.refetch();
-      void queryClient.invalidateQueries({ queryKey: queryKeys.settings });
-    },
-    onError: (e) => setStatus(`Google connect failed: ${e instanceof Error ? e.message : String(e)}`),
-  });
-  const dirtyGoogleClient = "google.client_id" in dirty || "google.client_secret" in dirty;
-
-  function discard() {
-    setDirty({});
-    setStatus("");
-  }
-
-  return (
-    <>
-      {/* Canonical sub-tab strip — above the panel card, shared with the rail
-          surfaces + plugin views (StageSubnav: single source of truth). */}
-      {categories.length > 1 ? (
-        <StageSubnav
-          active={category}
-          onSelect={setActiveCategory}
-          tabs={categories.map((c) => ({ id: c, label: c }))}
-        />
-      ) : null}
-      <section className="panel stage-panel settings-panel">
-      <PanelHeader
-        title="Settings"
-        kicker={dirtyKeys.length ? `${dirtyKeys.length} unsaved change${dirtyKeys.length === 1 ? "" : "s"}` : "applies on save"}
-        actions={
-          <>
-            <button className="secondary-button" type="button" onClick={() => testConn.mutate()} disabled={testConn.isPending || save.isPending}>
-              {testConn.isPending ? <Loader2 className="spin" size={15} /> : <ShieldCheck size={15} />}
-              Test connection
-            </button>
-            <button className="secondary-button" type="button" onClick={discard} disabled={save.isPending || !dirtyKeys.length}>
-              <RotateCcw size={15} />
-              Discard
-            </button>
-            <button className="primary-button" type="button" onClick={() => save.mutate()} disabled={save.isPending || !dirtyKeys.length}>
-              <Save size={16} />
-              Save &amp; apply
-            </button>
-          </>
-        }
-      />
-      <div className="stage-body">
-        {pendingRestart.length ? (
-          <div className="settings-banner" role="alert">
-            <AlertTriangle size={14} />
-            <span>Needs a restart to take effect: {pendingRestart.join(", ")}</span>
-          </div>
-        ) : null}
-        {status ? <p className="settings-status">{status}</p> : null}
-
-        {visibleGroups.map((group) => (
-          <section className="settings-group" key={group.section}>
-            <p className="settings-group-title">{group.section}</p>
-            {group.fields.map((field) => (
-              <SettingRow
-                key={field.key}
-                field={field}
-                dirty={field.key in dirty}
-                value={field.key in dirty ? dirty[field.key] : field.value}
-                onChange={(v) => setDirty((d) => ({ ...d, [field.key]: v }))}
-              />
-            ))}
-            {group.section === "Discord" ? (
-              <div className="settings-group-actions">
-                <button
-                  className="secondary-button"
-                  type="button"
-                  onClick={() => testDiscord.mutate()}
-                  disabled={testDiscord.isPending || save.isPending}
-                >
-                  {testDiscord.isPending ? <Loader2 className="spin" size={15} /> : <ShieldCheck size={15} />}
-                  Test connection
-                </button>
-                <a className="settings-help-link" href={DISCORD_GUIDE_URL} target="_blank" rel="noreferrer">
-                  How to create a bot <ExternalLink size={13} />
-                </a>
-              </div>
-            ) : null}
-            {group.section === "Google" ? (
-              <div className="settings-group-actions">
-                <button
-                  className="secondary-button"
-                  type="button"
-                  onClick={() => googleConnect.mutate()}
-                  disabled={googleConnect.isPending || save.isPending || dirtyGoogleClient}
-                  title={dirtyGoogleClient ? "Save the client ID + secret first" : undefined}
-                >
-                  {googleConnect.isPending ? <Loader2 className="spin" size={15} /> : <Link2 size={15} />}
-                  {googleStatus.data?.connected ? "Reconnect Google" : "Connect Google"}
-                </button>
-                <span className="settings-inline-status">
-                  {googleStatus.data?.connected
-                    ? `Connected${googleStatus.data.email ? ` as ${googleStatus.data.email}` : ""}`
-                    : dirtyGoogleClient
-                      ? "Save the client ID + secret, then connect"
-                      : "Not connected"}
-                </span>
-                <a className="settings-help-link" href={GOOGLE_GUIDE_URL} target="_blank" rel="noreferrer">
-                  Get an OAuth client <ExternalLink size={13} />
-                </a>
-              </div>
-            ) : null}
-            {/* Generic Test button (ADR 0029) — any plugin group that declares a
-                test endpoint (e.g. communication plugins). Discord/Google above are
-                bespoke and don't set `test`, so there's no overlap. */}
-            {group.test ? (
-              <div className="settings-group-actions">
-                <button
-                  className="secondary-button"
-                  type="button"
-                  onClick={() => {
-                    setTestingSection(group.section);
-                    testGroup.mutate({ endpoint: group.test!.endpoint, fields: groupFields(group) });
-                  }}
-                  disabled={(testGroup.isPending && testingSection === group.section) || save.isPending}
-                >
-                  {testGroup.isPending && testingSection === group.section ? (
-                    <Loader2 className="spin" size={15} />
-                  ) : (
-                    <ShieldCheck size={15} />
-                  )}
-                  Test connection
-                </button>
-              </div>
-            ) : null}
-          </section>
-        ))}
-
-        {/* Overview (read-only status snapshot + telemetry) — moved here from the
-            old Runtime section; not part of the settings schema. */}
-        {category === "Overview" ? <OverviewPanel /> : null}
-        {/* The delegate registry (ADR 0025) isn't part of the settings schema —
-            it's a CRUD surface, rendered as a custom panel under Integrations. */}
-        {category === "Integrations" ? <DelegatesSection /> : null}
-        {/* Plugin install/manage moved to its own top-level Plugins section. */}
-      </div>
-      </section>
-    </>
-  );
-}
-
-function SettingRow({
-  field,
-  value,
-  dirty,
-  onChange,
-}: {
-  field: SettingsField;
-  value: unknown;
-  dirty: boolean;
-  onChange: (value: unknown) => void;
-}) {
-  return (
-    <div className={`setting-row${dirty ? " dirty" : ""}`} data-key={field.key}>
-      <div className="setting-meta">
-        <label className="setting-label" htmlFor={`set-${field.key}`}>
-          {field.label}
-          {field.restart ? <span className="setting-restart">restart</span> : null}
-        </label>
-        {field.description ? <p className="setting-desc">{field.description}</p> : null}
-      </div>
-      <div className="setting-control">
-        <SettingInput field={field} value={value} onChange={onChange} />
-      </div>
-    </div>
-  );
-}
-
-function SettingInput({
-  field,
-  value,
-  onChange,
-}: {
-  field: SettingsField;
-  value: unknown;
-  onChange: (value: unknown) => void;
-}) {
-  const id = `set-${field.key}`;
-
-  if (field.type === "bool") {
+export function SettingsSurface({ tab = "overview" }: { tab?: SettingsTab }) {
+  if (tab === "telemetry") return <TelemetrySurface />;       // self-contained (own section + boundary)
+  if (tab === "plugins") {
     return (
-      <label className="setting-toggle">
-        <input
-          id={id}
-          type="checkbox"
-          checked={Boolean(value)}
-          onChange={(e) => onChange(e.target.checked)}
-        />
-        <span>{value ? "on" : "off"}</span>
-      </label>
-    );
-  }
-
-  if (field.type === "number") {
-    return (
-      <input
-        id={id}
-        className="setting-input"
-        type="number"
-        value={value === undefined || value === null ? "" : String(value)}
-        min={field.minimum}
-        max={field.maximum}
-        onChange={(e) => onChange(e.target.value === "" ? null : Number(e.target.value))}
+      <SettingsCategoryPanel
+        category="Plugins"
+        title="Plugin settings"
+        emptyHint="Plugins with their own view manage settings there. Anything view-less shows up here."
+        footer={<DelegatesSection />}
       />
     );
   }
-
-  if (field.type === "select" && field.options.length) {
-    return (
-      <select id={id} className="setting-input" value={String(value ?? "")} onChange={(e) => onChange(e.target.value)}>
-        {field.options.map((opt) => (
-          <option key={opt} value={opt}>
-            {opt}
-          </option>
-        ))}
-      </select>
-    );
-  }
-
-  if (field.type === "string_list") {
-    const text = Array.isArray(value) ? value.join("\n") : "";
-    return (
-      <textarea
-        id={id}
-        className="setting-input setting-textarea"
-        rows={3}
-        value={text}
-        placeholder="one per line"
-        onChange={(e) =>
-          onChange(e.target.value.split("\n").map((s) => s.trim()).filter(Boolean))
-        }
-      />
-    );
-  }
-
-  if (field.type === "secret") {
-    return (
-      <input
-        id={id}
-        className="setting-input"
-        type="password"
-        autoComplete="new-password"
-        value={typeof value === "string" ? value : ""}
-        placeholder={field.is_set ? "•••••••• (set — leave blank to keep)" : "unset"}
-        onChange={(e) => onChange(e.target.value)}
-      />
-    );
-  }
-
-  // string + select-without-options fallback
-  return (
-    <input
-      id={id}
-      className="setting-input"
-      type="text"
-      value={typeof value === "string" ? value : value === undefined || value === null ? "" : String(value)}
-      onChange={(e) => onChange(e.target.value)}
-    />
-  );
+  if (tab === "system") return <SettingsCategoryPanel category="System" title="System" />;
+  return <OverviewPanel />;                                    // overview (default; own section)
 }
