+++ tests/test_settings_schema.py
@@ -15,9 +15,9 @@ from graph.settings_schema import (
 def test_schema_groups_and_values():
     cfg = LangGraphConfig()
     groups = build_schema(cfg, model_options=["a", "b"])
-    # Grouped + ordered by category (ADR 0020): the Agent category leads, its
-    # sections in FIELDS order (Model, Routing, Identity).
-    assert [g["section"] for g in groups][:3] == ["Model", "Routing", "Identity"]
+    # Grouped + ordered by category: the Agent category leads, its sections in
+    # FIELDS order (Model, Routing, Goal mode, Tools, Identity).
+    assert [g["section"] for g in groups][:3] == ["Model", "Routing", "Goal mode"]
     fields = [f for g in groups for f in g["fields"]]
     # Every core FIELD is present. (build_schema also appends plugin-declared
     # settings — e.g. the discord plugin — so count only the core-keyed fields,
